+++ src/main/java/com/questland/handbook/service/model/hero/PrivateMultiplierDetails.java
@@ -8,7 +8,7 @@ import lombok.Data;
 // Ignoring unknowns to minimize internal API reliance
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class PrivateMultiplierDetails {
-  private int multiplier;
+  private double multiplier;
   private PrivateMultiplierLinkBonuses bonuses;
   @JsonProperty("spirit_bonus")
   private PrivateSpiritBonus spiritBonus;
